Reorder startup so we can set sys.stdin to None.
This fixes an ugly traceback if exit() (not sys.exit(), exit()) is
called.

exit() close sys.stdin if it can, and urwid's shutdown code tries to
use it. This fixes it by using urwid's support for starting and
stopping the screen before/after running its mainloop: we start the
screen, redirect std*, run the mainloop, then restore sys.std* before
stopping the screen. Now exit() does not have a real sys.stdin to close.

This mainly moves code around: the important part of the diff is
sys.stdin is now set to None.